PROBLEM 2 OF DAY 1
Imagine you are developing a program to track the history of web pages visited by a user in a web browser. Design a data structure using an array or a linked list to store the URLs in the order they were visited. How would you implement the functionality to allow the user to navigate back and forth through the history efficiently using stack-like behavior? Consider how you would handle cases where the user revisits a page or opens a new one while navigating through the history